Allow content to be only any block element
Currently when initializing ice's range the
content must have at least one p element.

In most of the cases it's ok, but sometimes we
want it to be only h1 or a list and this causes
a NOT_FOUND dom exception.

To initialize the range, we take the first block
element, not the first this.blockEl (which is p).