Fix for Issue 55: Webkit cut is broken
The problem stemmed from always inserting the icecut div into the main
document context. Now it uses the document context of the editor, which
is window.document for the contenteditable version and the iframe for
the TinyMCE version. This way Webkit does not get confused about what
element's contents to stick on the clipboard.