Suggestion: Update LICENSE / GPL2 or later
Strict GPLv2 is incompatible with AGPLv3 and I think may be incompatible with GPLv3 also. However, if you use the 'or later' clause as recommended by FSF, then you can auto-upgrade the terms of your license as the GPL is improved. In particular, GPLv3 is semi-compatible with AGPLv3 code, which would be a big improvement for the usability of ICE.